Update model definition and improve DB connection handling
This commit updates the weight attribute in the MyModel class from float to Decimal to enhance precision. Additionally, significant improvements have been made to how database connections are handled, establishing better logic for opening/closing connections and cursors. The database module changes also result in adjustments in main.py. Exception response messages in the Table class have also been made more descriptive to improve clarity of errors.

+++ models/ui_inputs.py
@@ -2,6 +2,7 @@ from pydantic import BaseModel
 from datetime import datetime, date
 from psycopg2 import sql
 from typing import Optional
+from decimal import Decimal
 
 
 class MyModel(BaseModel):
@@ -60,6 +61,6 @@ class Run(MyModel, BaseModel):
     invoice: int
     invoice_document: str
     waybill: str
-    weight: float
+    weight: Decimal
     date_departure: date
     date_arrival: Optional[date]
